controller без getList

+++ src/main/java/org/example/service/TaskService.java
@@ -3,11 +3,19 @@ package org.example.service;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.example.dto.NewTaskDto;
+import org.example.dto.TaskDto;
+import org.example.dto.UpdateTaskDto;
+import org.example.exeptions.ConflictException;
+import org.example.exeptions.StorageException;
 import org.modelmapper.ModelMapper;
 import org.example.model.Task;
 import org.example.repository.TaskRepository;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
@@ -18,25 +26,51 @@ public class TaskService {
 
     private TaskRepository taskRepository;
     private ModelMapper mapper;
+    private LocalDateTime time = LocalDateTime.now();
 
 
-    public Task createTask(Task task) {
-        return taskRepository.save(task);
+    public TaskDto createTask(NewTaskDto newTask) {
+        Task task = Task.builder()
+                .title(newTask.getTitle())
+                .description(newTask.getDescription())
+                .createdDateTime(time)
+                .deadline(newTask.getDeadline())
+                .status("New")
+                .assigneeId(newTask.getAssigneeId())
+                .authorId(newTask.getAuthorId())
+                .eventId(newTask.getEventId())
+                .build();
+        return mapper.map(taskRepository.save(task), TaskDto.class);
     }
 
-    public Task updateTask(Long id, Task task) {
-        return taskRepository.save(task);
+    public TaskDto updateTask(Long userId, Long id, UpdateTaskDto updateTask) {
+        Task task = taskRepository.findById(id).orElseThrow(() -> new StorageException("Событие не найдено или недоступно"));
+        if (!userId.equals(task.getAuthorId()) || !userId.equals(task.getAssigneeId())){
+            throw new ConflictException("Нет прав на изменение задачи");
+        }
+        task.setTitle(updateTask.getTitle());
+        task.setDescription(updateTask.getDescription());
+        task.setDeadline(updateTask.getDeadline());
+        task.setEventId(updateTask.getEventId());
+        return mapper.map(taskRepository.save(task), TaskDto.class);
     }
 
-    public Task getTask(long id) {
-        return taskRepository.findById(id).orElse(null);
+    public TaskDto getTask(long id) {
+        Task task = taskRepository.findById(id).orElseThrow(() -> new StorageException("Событие не найдено или недоступно"));
+        return mapper.map(task, TaskDto.class);
     }
 
-    public List<Task> getTasks() {
+    public List<TaskDto> getTasks(int page,int size, Long eventId, Long assigneeId, Long authorId) {
+        Pageable paged = PageRequest.of(page, size);
+
         return taskRepository.findAll();
     }
 
-    public void delete(long id) {
+    public void delete(Long userId, long id) {
+        Task task = taskRepository.findById(id).orElseThrow(() -> new StorageException("Событие не найдено или недоступно"));
+        if (!userId.equals(task.getAuthorId())){
+            throw new ConflictException("Нет прав на удаление задачи");
+        }
         taskRepository.deleteById(id);
     }
 }
